рефакторинг коду, розбиття по папкам
