+++ client/editor.html
@@ -79,6 +79,6 @@
         </div>
 
         <script src="./scripts/other/toogleThemes.js"></script>
-        <script src="./scripts/editor.js" type="module"></script>
+        <script src="./scripts/pages/editor.js" type="module"></script>
     </body>
 </html>
