+++ client/index.html
@@ -65,6 +65,6 @@
         <div id="toast" class="toast" aria-live="polite" hidden></div>
 
         <script src="./scripts/other/toogleThemes.js"></script>
-        <script src="./scripts/app.js" type="module"></script>
+        <script src="./scripts/pages/app.js" type="module"></script>
     </body>
 </html>
