+++ client/scripts/core/socket.js
@@ -1,0 +1,4 @@
+import { webSocketUrl } from "./paths.js";
+
+let socket = new WebSocket(webSocketUrl());
+export default socket;
