--- client/scripts/socket.js
@@ -1,4 +1,0 @@
-import { websocket_url } from "./paths.js";
-
-let socket = new WebSocket(websocket_url());
-export default socket;
